update unit tests, update 'VERSION'
