+++ src/syngen/VERSION
@@ -1,1 +1,1 @@
-0.8.0rc1
+0.8.0rc2
