+++ src/tests/unit/launchers/test_launch_train.py
@@ -1,13 +1,8 @@
 from unittest.mock import patch
 from click.testing import CliRunner
 
-from syngen.train import (
-    launch_train,
-    get_json_columns,
-    get_flattened_df
-)
+from syngen.train import launch_train
 from syngen.ml.worker import Worker
-from syngen.ml.data_loaders import DataLoader
 from tests.conftest import SUCCESSFUL_MESSAGE
 
 TABLE_NAME = "test_table"
@@ -31,7 +26,7 @@ def test_train_table_with_source_and_table_name(
     rp_logger.info(SUCCESSFUL_MESSAGE)
 
 
-@patch("syngen.train.preprocess_data")
+@patch("syngen.train.PreprocessHandler.run")
 @patch.object(Worker, "launch_train")
 @patch.object(Worker, "__attrs_post_init__")
 def test_train_table_with_metadata_path(
@@ -47,7 +42,7 @@ def test_train_table_with_metadata_path(
     rp_logger.info(SUCCESSFUL_MESSAGE)
 
 
-@patch("syngen.train.preprocess_data")
+@patch("syngen.train.PreprocessHandler.run")
 @patch.object(Worker, "launch_train")
 @patch.object(Worker, "__attrs_post_init__")
 @patch("syngen.train.setup_logger")
@@ -74,7 +69,7 @@ def test_train_table_with_metadata_path_and_source(
     rp_logger.info(SUCCESSFUL_MESSAGE)
 
 
-@patch("syngen.train.preprocess_data")
+@patch("syngen.train.PreprocessHandler.run")
 @patch.object(Worker, "launch_train")
 @patch.object(Worker, "__attrs_post_init__")
 @patch("syngen.train.setup_logger")
@@ -101,7 +96,7 @@ def test_train_table_with_metadata_path_and_table_name(
     rp_logger.info(SUCCESSFUL_MESSAGE)
 
 
-@patch("syngen.train.preprocess_data")
+@patch("syngen.train.PreprocessHandler.run")
 @patch.object(Worker, "launch_train")
 @patch.object(Worker, "__attrs_post_init__")
 @patch("syngen.train.setup_logger")
@@ -343,135 +338,4 @@ def test_train_table_with_invalid_batch_size(rp_logger):
         ["--batch_size", 0, "--table_name", TABLE_NAME, "--source", PATH_TO_TABLE],
     )
     assert result.exit_code == 2
-    rp_logger.info(SUCCESSFUL_MESSAGE)
-
-
-def test_get_json_columns_contained_one_json_column(rp_logger):
-    rp_logger.info(
-        "Test that the method 'get_json_columns' method "
-        "for the dataframe contained one json column"
-    )
-    path_to_data = "tests/unit/launchers/fixtures/data_with_one_json_column.csv"
-    data, schema = DataLoader(path_to_data).load_data()
-    assert get_json_columns(data) == ["_details"]
-
-
-def test_get_flattened_df_contained_one_json_column(rp_logger):
-    rp_logger.info(
-        "Test that the method 'get_flattened_df' method "
-        "for the dataframe contained one json column"
-    )
-    path_to_data = "tests/unit/launchers/fixtures/data_with_one_json_column.csv"
-    data, schema = DataLoader(path_to_data).load_data()
-    json_columns = ["_details"]
-    flattened_data, flattening_mapping = get_flattened_df(data, json_columns)
-    assert flattened_data.columns.to_list() == [
-        "id",
-        "created_at",
-        "updated_at",
-        "name",
-        "description",
-        "owner_id",
-        "is_default",
-        "is_encrypted",
-        "status",
-        "cluster_type",
-        "master",
-        "ssh_config",
-        "username",
-        "password",
-        "authentication_type",
-        "log_level",
-        "env_variables",
-        "key_passphrase",
-        "private_key"
-    ]
-    assert flattening_mapping == {
-        "_details": [
-            "cluster_type",
-            "master",
-            "ssh_config",
-            "username",
-            "password",
-            "authentication_type",
-            "log_level",
-            "env_variables",
-            "key_passphrase",
-            "private_key"
-        ]
-    }
-    rp_logger.info(SUCCESSFUL_MESSAGE)
-
-
-def test_get_json_columns_contained_more_than_one_json_column(rp_logger):
-    rp_logger.info(
-        "Test that the method 'get_json_columns' method "
-        "for the dataframe contained more than one json column"
-    )
-    path_to_data = "tests/unit/launchers/fixtures/data_with_two_json_columns.csv"
-    data, schema = DataLoader(path_to_data).load_data()
-    assert get_json_columns(data) == ["progress", "details"]
-
-
-def test_get_flattened_df_with_df_contained_more_than_one_json_column(rp_logger):
-    rp_logger.info(
-        "Test that the method 'get_flattened_df' method "
-        "for the dataframe contained more than one json column"
-    )
-    path_to_data = "tests/unit/launchers/fixtures/data_with_two_json_columns.csv"
-    data, schema = DataLoader(path_to_data).load_data()
-    json_columns = ["progress", "details"]
-    flattened_data, flattening_mapping = get_flattened_df(data, json_columns)
-    assert flattened_data.columns.to_list() == [
-        "id",
-        "created_at",
-        "updated_at",
-        "start_time",
-        "finish_time",
-        "status",
-        "pipeline_id",
-        "user_id",
-        "info.finished",
-        "info.total",
-        "step",
-        "description",
-        "source.id",
-        "source.name",
-        "source.connection_string",
-        "target.id",
-        "target.name",
-        "target.connection_string",
-        "cluster.id",
-        "cluster.name",
-        "cluster.master_webui",
-        "cluster.history_server",
-        "integrity_type",
-        "total_tables",
-        "included_tables",
-        "base_table"
-    ]
-    assert flattening_mapping == {
-        "progress": [
-            "info.finished",
-            "info.total",
-            "step"
-        ],
-        "details": [
-            "description",
-            "source.id",
-            "source.name",
-            "source.connection_string",
-            "target.id",
-            "target.name",
-            "target.connection_string",
-            "cluster.id",
-            "cluster.name",
-            "cluster.master_webui",
-            "cluster.history_server",
-            "integrity_type",
-            "total_tables",
-            "included_tables",
-            "base_table"
-        ]
-    }
     rp_logger.info(SUCCESSFUL_MESSAGE)
