+++ src/tests/unit/preprocess/fixtures/metadata_for_table_with_two_json_columns.yaml
@@ -1,0 +1,3 @@
+test_table:
+  train_settings:
+    source: tests/unit/launchers/fixtures/data_with_two_json_columns.csv
